pyproject.toml
[build-system]
#requires = ["setuptools >= 61.0"]
#build-backend = "setuptools.build_meta"
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/telomap/version.py"

[project]
name = "telomap"
dynamic = ["version"]
dependencies = [
  "pandas>=1.4.2", 
  "numpy>=1.22.4", 
  "scipy>=1.8.1", 
  "matplotlib>=3.5.2",
  "pysam>=0.19.0", 
  "seaborn>=0.11.2", 
  "natsort>=8.1.0", 
  "biopython>=1.79", 
  "scikit-learn>=1.0.2",
]
requires-python = ">= 3.8"
authors = [
  {name = "CY Tham", email = "chengyong.tham@u.nus.edu"},
]
description = "A tool for analyzing telomeres in telobait-captured or WGS long-read sequencing data"
readme = "README.md"
classifiers=[
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3.8",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords=["telomap", "telomere", "pacbio", "ont", "oxford nanopore", "long read", "WGS"]

[project.urls]
Homepage = "https://github.com/cytham/telomap"
Repository = "https://github.com/cytham/telomap.git"
Issues = "https://github.com/cytham/telomap/issues"
Changelog = "https://github.com/cytham/telomap/blob/main/CHANGELOG.txt"

[project.scripts]
telomap = "telomap.telomap:main"

